Remove old config folder
There is no need to keep the old folder around. I haven't referenced it
in some time. Also, it remains in the history of this git repo, so I
don't need to keep a copy of it.

Finally, I'll be reworking my nvim config by swapping over to TJ's
kickstart.nvim config and making modifications as needed.